world: clarify _spawn_character default eval time
Use it explicitly as it's not exactly clear at which state the appearance
is taken (on script initialization? on _init? on function call?)

I think I've used this syntax in Python (and Luau? forgot) but not with
globals/statics that may change

The only reference for GDScript I found was this, but it doesn't clearly
state when the evaluation takes place (or even mention it exists):

https://docs.godotengine.org/en/stable/tutorials/scripting/gdscript/gdscript_basics.html#variadic-functions

Also reformats the base class